Added cell splitting logic in tracking module
Updated labels_graph to a subclass of networkx.DiGraph, where nodes are
a Node namedtuple of (time, label). Has methods to add nodes (and edges)
from labeled images, and a constructor to build the graph from a stack
of labeled images.

Function split_nodes iterates through merged or splitting nodes in the
graph, and uses the temporal information as a starting point for the
watershed algorithm.

Signed-off-by: Mauro Silberberg <maurosilber@gmail.com>